global: reformat code with black

+++ invenio_app_ils/patrons/anonymization.py
@@ -23,10 +23,7 @@ from invenio_app_ils.circulation.search import (
     get_active_loans_by_patron_pid,
     get_loans_by_patron_pid,
 )
-from invenio_app_ils.errors import (
-    AnonymizationActiveLoansError,
-    PatronNotFoundError,
-)
+from invenio_app_ils.errors import AnonymizationActiveLoansError, PatronNotFoundError
 from invenio_app_ils.ill.proxies import current_ils_ill
 from invenio_app_ils.notifications.models import NotificationsLogs
 from invenio_app_ils.patrons.api import get_patron_or_unknown_dump
@@ -97,9 +94,7 @@ def anonymize_patron_data(patron_pid, force=False):
             )
         )
     OrderSearch = current_ils_acq.order_search_cls
-    n_orders = (
-        OrderSearch().get_ongoing_orders_by_patron_pid(patron_pid).count()
-    )
+    n_orders = OrderSearch().get_ongoing_orders_by_patron_pid(patron_pid).count()
     if n_orders > 0:
         raise AnonymizationActiveLoansError(
             "Cannot delete user {0}: found {1} active orders.".format(
@@ -210,17 +205,12 @@ def delete_user_account(patron_pid):
         dropped += d.delete(synchronize_session=False) or 0
 
         dropped += (
-            SessionActivity.query.filter(
-                SessionActivity.user_id == patron_pid
-            ).delete()
+            SessionActivity.query.filter(SessionActivity.user_id == patron_pid).delete()
             or 0
         )
 
         dropped += (
-            UserIdentity.query.filter(
-                UserIdentity.id_user == patron_pid
-            ).delete()
-            or 0
+            UserIdentity.query.filter(UserIdentity.id_user == patron_pid).delete() or 0
         )
 
         ra = RemoteAccount.query.filter(
@@ -236,10 +226,7 @@ def delete_user_account(patron_pid):
             dropped += ra.delete() or 0
 
         dropped += (
-            UserProfile.query.filter(
-                UserProfile.user_id == patron_pid
-            ).delete()
-            or 0
+            UserProfile.query.filter(UserProfile.user_id == patron_pid).delete() or 0
         )
         dropped += User.query.filter(User.id == patron_pid).delete() or 0
 
@@ -253,7 +240,8 @@ def anonymize_patron_in_notification_logs(patron_pid):
 
     with db.session.begin_nested():
         notifications = NotificationsLogs.query.filter_by(
-            recipient_user_id=patron_pid).all()
+            recipient_user_id=patron_pid
+        ).all()
         for notification in notifications:
             notification.recipient_user_id = AnonymousPatron.id
             notification_anonymizations += 1
